+++ pymince/file.py
@@ -96,21 +96,22 @@ def replace_extension(filename, old_ext=None, new_ext=None):
         return filename
 
 
-def decompress(src_path, dst_path):
+def decompress(src_path, dst_path, size=64 * 1024):
     """
     Decompress given file in blocks using gzip.
 
     :param str src_path: source file path
     :param str dst_path: destination file(unzipped) path
+    :param int size: Read up to size bytes from src_path for each block.
     :return: dst_path
 
      Examples:
         from pymince.file import decompress
 
-        decompress("/foo/src.txt.gz", "/baz/dst.txt") # --> "/baz/dst.txt"
+        decompress("/foo/src.txt.gz", "/baz/dst.txt")  # --> "/baz/dst.txt"
     """
 
     with gzip.open(src_path) as src, open(dst_path, mode="wb") as dst:
-        lines = iter(functools.partial(src.read, 64 * 1024), b"")
+        lines = iter(functools.partial(src.read, size), b"")
         dst.writelines(lines)
     return dst_path