+++ tests/test_algorithm/test_fibonacci.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
-import pytest
 
 import pymince.algorithm
 
 
 def test_fibonacci_n_lt_0():
-    with pytest.raises(ValueError):
-        pymince.algorithm.fibonacci(n=-1)
+    expected = ()
+    res = pymince.algorithm.fibonacci(n=-1)
+    assert tuple(res) == expected
 
 
 def test_fibonacci_n_eq_0():
@@ -16,6 +16,12 @@ def test_fibonacci_n_eq_0():
 
 
 def test_fibonacci_n_eq_10():
-    expected = (0, 1, 1, 2, 3, 5, 8, 13, 21, 34)
+    expected = (0, 1, 1, 2, 3, 5, 8)
     res = pymince.algorithm.fibonacci(n=10)
+    assert tuple(res) == expected
+
+
+def test_fibonacci_float_n():
+    expected = (0, 1, 1, 2, 3, 5, 8)
+    res = pymince.algorithm.fibonacci(n=10.4)
     assert tuple(res) == expected